Fix open graph / twitter scheme
Only in development would we be serving on http.  In production we
serve http but proxied by an nginx instance that is terminating an
https connection.  So we may not perceive that we are talking https,
but, where it matters, we are.  Therefore, don't be clever: just say
https always and deal with it in development if necessary.

Part of #15.