Add first draft of image generation function
The purpose of this, which is happening too late to be used in the
Nantes municipal elections, is to solidify and write down what I've
learned about pillow and to remind me how to do a minimal working
example of what I eventually want.

This also adds the Montserrat font that Transport Nantes uses in its
logo.

Part of #45.