Make newsletter list configurable
We now present a checkbox list (inclusion list) of available
newsletters and let the user choose to which to subscribe.

The form doesn't know anything about the user's actual subscriptions,
and, indeed, that's not stored in a particularly accessible way.  I'll
want to cache the last known state somewhere.

Part of #69.