Provide mailing list signup that is site-friendly
It was getting confused about base templates on links (url resolution)
in templates.  The only way I found to make this work was to derive
inherited views so that the functionality stays the same between
different sites but I can parametrise the pages differently.

The disadvantage is I'll need to reproduce all the *_m files again for
each new site.  So this won't scale, but it gets me out of a mess today.

Part of #95.