Add petition template for four frame petitions
This lets us present the petition text in four separate sections that
we can style independently.  Here we add the second section in blue.
Eventually this should be more configurable.

Add the signing form at the bottom as well except on very small
screens (xs) where it only appears at the bottom.