Add dashboard app and a first dashboard
We add a dashboard to summarise list and petition memberships.

I haven't sorted user permissions yet on this view.  I must do that
soon, but it's not working yet, and I'd like to get this out.  The
data isn't (currently) sensitive, should someone stumble on the view,
and it doesn't permit any interaction.