Fix CSRF error
Sometimes the CSRF token was not present in the request cookie
(presumably, on a first visit), and then I'd 500.

I think what's going on is that it's not present at first visit.  This
will require additional exploration to make sure the forms I want to
render have their csrf tokens.