Remove old post-election observatory
This app was incompatible with reality.  It required heavy citizen
involvement on a subject that most citizens don't much care about.  It
didn't significantly further the goals of the organisation.  And
elected officials hve zero interest in reporting their intended
progress outside of the heady and untested waters of elections.

Part of #121.