Reversed the order of appearance of the two forms.
Also merged main.js which handled Stripe call with
form_flow.js.

Previously the validation of donation_form was server side, but dynamic
transformations of required fields made it difficult to
validate, so used the more server-trip friendly option of
HTML5 to perform validation.

Amount form wasn't validated, but user wouldn't notice,
Strip would just not initiate a checkout session without
proper informations (aka amount or subscription).
Now HTML5 does process the form and displays errors when
appropriate.

The function in views.py was called to validate the donation_form,
as this is now done via the browser, this function is uneeded.

Part of #54