Add grouped layers to toggle display.
Base layers are now "groups" which have sub-layers ("subgroup").
When you toggle the group, all subgroups are toggled as well.

On the front size, the layers do disappear, but layer boxes aren't
unchecked.

You still can toggle subgroups individually.

This change is meant to make it easier to filter out some parts for
users.

Part of #118