Add View to retrieve and save tracking information on server.
Using Ajax, we retrieve a csrf-controled set of 2 bool values
illustrating the progress of user in the form.
This request is done only when the user leaves the page.

We then use the state of the booleans to build an entry in our
tracking model, saving progress and timestamp.

Part of #54