Add tracking in the Javascript of donation page.
Tracking is done using 2 booleans :
- step_1_completed : false by default
- step_2_completed : false by default

When user press continue on the 1st step, a form validation is
made by HTML5. If this validation returns true, the step_1_completed
bool is set to "true" and the 2nd form is displayed.
Same process occurs for 2nd step with personal information form,
but validation launches the rediraction to appropriate checkout session.

The redirection or the closure of the window containing the tab
triggers the beforeunload event, and state of booleans is sent to the
server alongside with csrf token to make sure the POST request comes
from browser.
Using the body of this request, we build and entry to database to track
and save progress of users.

Part of #54