Add free amount handling to form_flow.js
The goal here is to handle the specific case where user selected
free amount and ajust required fields accordingly.

By default, the value associated with "Montant Libre" field is 0.

- payment_amount_selected() checks what amount is selected among
the amount fields, and returns its value.
- An Event listenner is added to all single time payments radio buttons
and checks if the value is 0 (the value of free amount button)
If the value is 0 then the text field becomes required, if another
button is clicked, then it's not.