Add progression markers
There are two ways to track progres, one for user and one for admins.

The two boolean var "step_[int]_completed" are used to track progression
for admins, the value of those bools will later be retrieved to
know how far the user went.

The user cue is the progress bar on top of the form.
To toggle the progress of this bar, we add class "active" to the
corresponding item in the progress bar.
For instance, each step on the form is identified with an id like :
"progression_bar_step_[int from 1 to 3 included]".

In this commit, we add 2 events listeners that fire once the user
clicks their respective "Continue" button.

The event listener attached to the amount form, that fires on submit,
will first check for the form validity. If the form is indeed valid,
it hides the amount form and displays the donation form asking for
personal informations.
It also toggle step_1_completed to true and add the 'active' class to
the corresponding item in the progress bar, lighting it up for the user.

The 2nd event listener attached to the personal form, that fires on
submit, will check for the form validity. If the form is indeed valid,
it will toggle the step_2_completed to true.
Further instructions are attached to donation form introducing Stripe,
those instructions will be in another part of the script regarding
the buttons.

Part of #54