Add a function to create a Stripe checkout session
Creating a checkout session creates a link the user is redirected to
at the end of the form once all information is valid.
This link host a checkout session where the user can enter its credit
card information and pay the donation he did set up.

The checkout session creation process requires different arguments
depending on the type of donation : One time payment or subscription.

- Both type need :
  - Valid success URL and cancel URL to which the user will be
    redirected after the payment.
  - A payment method, set to card but more exist.
  - mode : Type of donation
  - A currency, set to EUR.
  - customer email, retrieved from the form.
  - A quantity.

- One time payment specifically need:
  - A name for the "product" sold, in our case "Donation"
  - An amount in cents. (5€ = 500) This amount is input as normal
  euros in the form, we then convert it to cents using the
  order_amount(items) function.

- Subscription specifically need:
  - A "price" which is a Stripe ID for a product we created
  beforehand in the Stripe dashboard. In this product we set
  both the amount and the recurrence period. (Every X months)
  This is the only way to have susbscription plan, as for today
  users can't create a custom amount for a subscription, and are
  limited to the options we give them.

Alongside the mendatory information, we also pass in the arguments
metadata, which is a dictionary of key-value pairs. The metadata
contains all information gathered in the form.
Once the checkout session completed, Stripe will send events confirming
the payment to the webhook we will setup, and we're going to use
metadata to create a Donation entry in our database.

Part of #54