Add progression tracking in View and JS
The progress is measured with two bool variables:
"step_1_completed" and "step_2_completed".

They default to false.
The first variable is set to "true" when the user completes a valid
amount form and clicks "Continuer".
The second variable is set to "true" when the user has provided
valid donation form and clicks "Soutenir".

When the user leaves the donation page, whether it be because it's
redirected to Stripe or abandons the donation process, the
"beforeunload" event is triggered by the browser.

On the beforeunload event, the "step_1_completed" and "step_2_completed"
are both sent to a new url at donation/tracking/ where the function
"tracking_progression" is called.

This function will take the body of the request and check if the
values of both "step_1_completed" and "step_2_completed".
Because in the process of sending the request, the values are
strings containing "true" or "false", the function will convert
them to booleans.

Once the values are converted, we add a timestamp and create a new
object of the class TrackinProgression which will inform us
about the progression of the user.

We will be able to see the abandons and success from the admin panel.

Part of #54