Add Stripe Checkout redirection
This part of the code triggers when the user submits the donation form.
If the form is valid, we send the content of all fields to the server
which will create and return to Javascript a Session ID, if all went
well in the Checkout creation process in views.

The session ID is used to redirect user to the appropriate page,
where it will be able to complete the donation by paying.

Part of #54