Add Stripe webhook management
Stripe can send to a webhook diverse events to inform our system about
what's happening in the Stripe checkout.
The event in charge of validating that the payment went through is the
"checkout.session.completed" event.

In the stripe_webhook function, we will receive the event and we will
validate that the event is the "checkout.session.completed" event.

If the event is the "checkout.session.completed" event, we will
create a Donation entry in the database.

To do so, we extract the metadata that we previously sent to Stripe
to pass them as parameters for the Donation entry.

One of the arguments is a username.
This username is randomly generated by the get_user function, using the
email to look if the user already exists.
Doing so, if a user wants to change its email, we can edit the
email on its user account.

Once all information is extracted, we create the Donation entry
that will allow us to produce a tax receipt to send to the donor.

Part of #54