Add TopicBlogItem Model
TopicBlogItem is a model that represents a topic blog item.
It can be an article, a newsletter or any other kind of content.

It holds every possible field that can be used to create content.
Fields in the future are going to be filtered to display only the
relevant ones for the current content type.

Part of #115