Add TopicBlogTemplate Model
This model will list all templates available to create various
sorts of content.
For now, its template_name attribute points to an existing template
in the templates directory.
It has to be in the format "topicblog/name_of_template.html"

The comment field explains the intent of this template.

Part of #115