Remove large crowdfunding banner and display simple donation request
The crowdfunding banner with its minimum height and teasing close
button that just opens again on the next click was driving me insane.
I have to assume it would do the same to real users.

The large banner might have worked but (1) in tests it clearly didn't
and (2) the close button would have to be persistent, and that's a
level of complexity we shouldn't implement at this stage.

This commit remove them both and replaces with a simple request.