Add topicblog base template and basic control panel
This permits us to provide a control panel / banner for permissioned
users and so avoids having to pass through various ancillary pages.

Rename function set_context() to set_social_context() so that we don't
need to comment each usage about what it is.

Simplify the new/edit TBItem view class.

Part of #326.