Record session id correctly in utm
I was using the wrong session id name and reading it from the wrong
place.  That this worked sometimes was an anomoly due to bad testing,
probably because I was logged in when testing.  And simple bad testing.

In addition, fix the session cookie middleware, which was only adding
cookies at response time.

Add test to catch this in the future.

Part of #347.