Rework login workflow
The login workflow was largely broken with large chunks unimplemented.
The form displays were rudimentary and unintelligible.  This commit
takes a first stab at rectifying that.

This is a pedagogical commit, lacking tests and surely containing a
handful of bugs.  The point is to hand off to @shriukan33 to do the
tests and so to learn how this code works.

Part of #51, #119.