Remove item_sort_key
A good idea that we turned out not to like using.  It also became
impractical.  For example, using it to A/B test social media content
left open doubt on click-through to the site, what social media tags
had been shown to the user.  We'll have to do A/B testing through
another means.

At the same time, and to support that change, remove the servable
field.

Somewhat opportunistically, introduce a first_published_date and
retire the never-used header_slug.  I'd initially anticipated using
header_slug to allow click-through on image banners, but @shriukan33
convinced me that click-through on the banner would make more sense.
Since that conversation, we've realised we don't do that, either,
because the article we are presenting is immediately below the image.
So clean up here.

Explicitly out of scope in this change:
  * There's vast repetion between the three test files
    tests_{auth,unauth,staff}_users.py.
  * The tests are not complete.  For example, publication is not
    tested.
  * The tests don't quite capture the right idea of making sure only
    one item with a given slug is published at once.

Closes #371.  Closes #397.  Part of #414.