Deprecate ContentType and TBTemplate
In prevision of further changes, we're deprecating the ContentType
and TBTemplate.
Their features are now transfered to TBItem as attributes.

Every reference to ContentType and TBTemplate have been removed except
the models. Because the table isn't deleted in the database, leaving
the model leaves a trace of the reason of their existences.

Part of #455 #429