Fiddle with sort order
For choosing the publishable page, it doesn't matter a wit in what
order things were modified, only in what order they were published.
So only sort on publication date.

For the list view, it turns out to get confusing when including
publication date in sort order.  Our heads mostly remember when we
modified things, and more recent things should come first.  So do
that.

Part of #462.