Update dashboard to know about TBEmail
This commit doesn't concern itself with pretty presentation.  Tweaking
and fine-tuning the visuals is left for another commit (though
ideally before branch merge).

Indeed, it would probably be good to present a CSS table with the
different classes of object we support and links to create or list
each object type, thus creating a more uniform interface.

Closes #526.