Add base templates to render TBEmails in mail clients
We're going to use 2 sorts of templates to display TBEmails:
- One to be displayed in mail clients, which heavily relies on inline
styles and HTML tables
- One to be displayed on our website, with a more article looking way,
that will allow the user to browse the website normally.

This commit introduces the first kind of template.

Along with that, email_tags are reusable blocks that allows us to
modulate the layout in an easier way.

Note : The custom markdown tags we use for TBItems might not display as
expected in mail clients because they don't support most of modern
style features (No JS or bootstrap), but they will display normally
on our website, as they use a different template.

Part of #532