Add topicblog launcher edit and add page
On model :

Move the function get_image_fields
from TopicBlogItem to TopicBlogOjectBase
now all child class of TopicBlogOjectBase
have access to the get_image_fields function
(this function is necessary for form_post_process)

Add template_config in TopicBlogLauncher
template_config is a dictionary that list
all available template for launcher with their
fields requirement.

On view :

Add form_post_process in TopicBlogLauncherEdit
for keep the image field when we edit a
TopicBlogLauncher object.

on form :

Create the TopicBlogLauncherForm

Part #531