Fix content_launcher.html
The template topicblog/content_launcher.html, which we use to display
TBLauncher objects in test / preview mode, contains a link to whatever
its object article (TBItem) is.  It's completely reasonable, however,
that a TBLauncher starts life without an article_slug.  This is
validated at publish time.

Without the article slug, however, url reverse doesn't have an
argument and so doesn't match.

In order to prevent display error, then, don't display the link if it
hasn't yet been provided.

Note that (1) the publish workflow must validate that all necessary
components are present and (2) TBLaunchers are displayed via
templatetag as links to our projects, never independently, solo, using
this template.

Part of #528 and #531.