Enable easier mailinglist signup for authenticated users
If the user is authenticated, we needn't ask for email address or
present a captcha.  Otherwise, ask for email and require captcha.

In addition, remove kludge in which we checked form inheritance to
decide in what case we found ourselves.  Instead, just read the fields
we expect from the form and behave accordingly.

Part of #383.

Commented out failing tests (see #560)
Some parts of the views didn't work as intended.