Improve timed token security
Our timed tokens were emminently forgeable, although it seemed quite a
lot of work for an adversary to do so for the very little gain.  This
commit addresses that issue by upping (significantly) the amount of
work necessary to forge a timed token.  We believe the difficulty of
forging a timed token is now on a par with the difficulty of breaking
SHA1 or SHA256.

This increases the length of our timed tokens, but we don't currently
consider that to be a problem.