Fix unproper synthax used on queryset
Querysets can be empty, if we want to get the first object of a queryset
or None, we must use .first() method that wil return None if the queryset
is empty.
Trying to get the first object of an empty queryset will raise an error
if done with queryset[0] : out of range.