Adjust ViewOne template to delete unused buttons and add another
Retirer et Republier buttons never were implemented and are outdated in
our current design, so we are removing them.

Added a button that allow the user to see the currently published
version of a slug if any, even from drafts.
It does 404 on click if no slug is published, but it doesn't seem to be
a big issue for now.