Add test for the new index page
This test merely confirms that something loads successfully at '/'.
It's not what we'll eventually need, but it will remind us to do
better.

We are staging the new index page at /index2 until we're happy with
it.  Note that we'll have to update tests again when we make that
change.

Part of #581.