Make the number of displayed words adjustable in item-teaser
It's a proposition to make the number of words displayed adjustable
depending on what we want to display.
I think it could be used to show just enough to tease curisity from the
user, but that line depends on the item we apply the teasing to.

For now it's hardcoded, but so is the item so it's still a git action to
edit this value.
However, I think we could add an optional field "teasing_length" to the
TBItem model to make it configurable in the future.