Fix image banners overlapping the Navbar
In commit 70d41a8262a67edb5437e49e9098d1e67f84e614, the ul tag was
removed from a div separating the navbar from the content.
The UL tag inside it forced a small gap that was then covered by a
negative top margin ("mt-n3"). The removal of this tag made the negative
margin overlap the navbar, because it didn't have the ul space to cover.

Instead of messing with the CSS, I removed the useless ul tag alongside
its parent div (that was also useless because it had no properties).
Removing this div also removes the need to have a negative margin that
covers the gap left by the ul tag.

This negative margin was present in base_mobilitain.html but also in
most base templates, because they came from base_mobilitain.html.

Closes #599