Update item_teaser test
The previous iteration of item_teaser was no able to format properly
when used in combination with the `tn_markdown` templatetag.
This use made us use a div instead of a p tag to wrap the teaser's text.

This change had to be reflected in the test that verifies that the
output has a div with 'teaser-text' class wrapping the inserted text.