Refactor and improvement of TopicBlog's List views
Why a refactor:
All the list views have the same implementation, with as the only
difference the urls it uses in the template. This led to the creation
of multiple templates that were extremely similar and any change had to
be duplicated to every template.

How did I refactor:
I created a base template that is used by all the list views.
To populate it with the correct urls, I added to the model the
url name that is set inside urls.py, and accessed these properties
inside the base list view.

About the improvement :
Added a Font-Awesome + sign on every list template as well as an arrow
to get back either to the general list or to the Dashboard.
This allows a smoother navigation into the CMS by providing shortcuts
to the various parts of the dashboard.