Add TopicBlogMailingListPitch model
Mailist list pitch is a page that promotes a mailing
list to the user, in order to get them to subscribe.

This model sets the fields for the mailist list pitches.

Because people come in different shapes and sizes, we want to offer
multiple reasons to subscribe to the list, depending on their
personality. Creating Mailing list pitches gives us the option to have
dedicated pages to promote each list, using various arguments.

Part of #637