Improve TBEdit form's error displaying
The errors used to be displayed only in new items, and not while
editing an exisiting item.
The reason was that we overrode the kwargs["form"] which may contain
the form with errors whith a form that didn't have errors.

On the style side, the errors were displayed in black and above the
field, while it's widely accepted that the errors should be displayed
in red and below the field.