Add Django-ses logger
The lib uses the getLogger(__name__) method to get a logger instance.
__name__ being evaluated as the module name, and editing a module's code
being out of option, one way to display its output to our logs is to
add to django a django-ses logger.

Because the extra parameter isn't logged by Django's default logger,
we need to create a new handler to get the full details.

Part of #714