Measure teaser length in words rather than characters
Words is a more logical and easily intuitable measure for most humans,
and we would always like to cut on word boundaries.

Both cutting at words and cutting at characters run the risk of
breaking in the middle of markdown or html.  The eventual fix for this
is that we should incorporate this functionality into the markdown
parser, with possibility that it doesn't cut exactly so as not to clip
a multi-word link, for example.

Closes #730.