Measure teaser length in characters rather than words
Whilst words may be a bit easier to think about than characters for
truncating teasers, but it created more problems than it solved.  Most
notably, truncate_chars preserves newlines but truncate_words does
not, which meant that a TBTeaser on a TBItem that begins with a '#'
rendering its entire content in a header tag.

The ultimate solution is to push the truncation into tn_markdown.

This rolls back ef739ca6ed, #730.