Make LoginView tests use TransactionTestCase
The error raised was :
raise TransactionManagementError
( django.db.transaction.TransactionManagementError: An error occurred
in the current transaction. You can't execute queries until the end of
the 'atomic' block.)
(See full log here :
https://github.com/transport-nantes/tn_web/runs/6557513431?check_suite_focus=true)

After googling this error I found this StackOverflow topic :
https://stackoverflow.com/questions/43978468/django-test-transactionmanagementerror-you-cant-execute-queries-until-the-end
And this one :
https://stackoverflow.com/questions/21458387/transactionmanagementerror-you-cant-execute-queries-until-the-end-of-the-atom#comment44444356_23326971

Both mentioned that switching testCase to TransactionTestCase worked,
and it did for me as well.

Because the difference between testCase and TTCase is slim and doesn't
change anything in production code, I decided to not try to dig inside
database locks in test context.
A fix for this is the use TransactionTestCase, which according to docs,
are almost identical :
TransactionTestCase and TestCase are identical except for the manner in
which the database is reset to a known state and the ability for test
code to test the effects of commit and rollback:

    * A TransactionTestCase resets the database after the test runs by
    truncating all tables. A TransactionTestCase may call commit and
    rollback and observe the effects of these calls on the database.

    * A TestCase, on the other hand, does not truncate tables after a
    test. Instead, it encloses the test code in a database transaction
    that is rolled back at the end of the test. This guarantees that
    the rollback at the end of the test restores the database to its
    initial state.

see :
https://docs.djangoproject.com/en/4.0/topics/testing/tools/#transactiontestcase