Make tbi_slug templatetag context-aware so that it can render mail
We need to render email differently than web pages: our CSS isn't
available and relative URLs fail.  Client-side render context is
different.

This commit is a first step: it makes the function tbi_slug
context-aware so that it can resolve an FQDN for the otherwise
relative URl.

Part o f#547.