Capture and log click and open signals from SES
We catch the click and open signals from SES and log them, so that
if we someday change our mind about using these features, the code is
already there.
It also tells future reader that we didn't forget to handle these,
simply we have reasons to not use them at this moment in time.

Closes #778