Add Unique constraint to prevent multiple sends to same recipient
If one is subbed to multiple newsletters and we want to send the same
article to the those newsletters, we need to make sure that we don't
send it to the same person more than once.
In the event we try to do so, we log the error and move on to next
recipient.

Closes #752