Allow Admins to send themselves same slug multiple times
As a convenience, and for admins only, an admin can send to themselves
and to other admins the same article multiple times.

It's done by simply appending a random string to the slug to not violate
the Unique constraint.

The Send record is created with a modified slug, but keeps its features:
open times, click times, etc because all those take the ID into
consideration, not the slug.

This is a quality of life improvement for admins.