Add missing fields and reorganize settings_local dev template
Some settings were missing (Like AWS configuration set name) compared
to infrastructure's settings.
I reorganized the settings orders to fit the order used in infra.

I also left the BASE_DIR because although it is not used in infra,
it's useful for devs that might use this template so their db is in the
right place.

Closes #789