Make Mobilito PWA-ready
We want users to have an easy access to this app, and looking for it
through their browser is a brake to adoption.

This allows users to download the app from their browser and add a
shortcut to their home screen. It looks like a native app, but is really
just a web app.
See :
https://developer.mozilla.org/fr/docs/Web/Progressive_web_apps

Part of #40