Change date_modified to date_created
This is essentially what we were using it for, since we don't permit
modification of TBObjects.  Since the previous commit, we also were
only setting the timestamp on date_modified (now date_created) on
creation.  So this commit simply changes the variable name everywhere
to bring it into alignment with usage.

This commit should have no functional impact.

Closes #725.