Make selected templates rely on the displayed object
As explained in #862, templates were hardcoded.

In the current build, all TB objects are displayed using their
.template_name attribute. The problem is, in TBE and TBP, the
.template_name attribute is set to the email version ! It only
displayed right on web because the view was overriding
self.template_name with an hardcoded value.

The new behavior will always use the object's .template_name attribute
for web, and will look at its email_template counterpart in the
template_config dict to produce emails.

However, it doesn't change the fact that all previously created objects
are still using the email version as their template_name attribute, and
trying to display them on web wont work.

To fix this, we added a data migration that replaces all the
incorrect template_name attributes with the correct ones.

Closes #862