DROP ME
In this commit I gave a go to implement a ViewOne version, only
to realize, once it was done, that I found it pointless.

This ViewOne version shows the _wrapped object's_ ViewOne version,
while we would like to show something that makes sense like a preview
of the social media sharing.

However I couldn't find any offline tool to achieve that goal yet.
Some websites offer a preview for publicly accessible URLs, but
the point here is precisly to preview before going public.