Allow user to point to a position on the map to provide device location
We used to localise the user on clicking the
"Je partage ma position" button. This was not ideal because
it introduced an unuseful extra step (revealing the map)

and because the user should be able to position himself
on the map without triggering geolocalisation.

With this commit the map is centred on Nantes at wide
(departement) zoom. The user geolocalises only on clicking
the location control.

Part of #932.