Add session_sha1 field to Mobilito's session model
We compute the session's SHA1 and save it to later be able to access it
through a URL (e.g. /mobilito/session/<SHA1>).
Because the sha1 is meant to be an unique field, we need to make 3
different migrations to add the field, populate it, and then make it
unique.

This migration does the first step by adding the field.

Part of #927.