Create flag_session view
This view creates a flag on a session, making sure that the
flag as been only done once.
It adds to logs the flag of a session, and creates an InappropriateFlag.

This is only available for auth users, as we don't want to allow
Joe Random to spam flags.

Part of #928