Move image field validation to publication instead of saving
This commit allow the TBObjects to keep the images on edit.

One should be able to save an unfinished work, but not publish it.
The clean() method used to enforce the header image validation leading
to a form_invalid() call when saving an unfinished work.
This would prevent the images from persisting, as we retrieve the images
in the form_valid() method.

Closes #922 #921