Improve navbar's hamburger responsive and add Petite Mobilite link
The navbar progressively displays fewer elements instead of hiding
everything under a hamburger menu at 992px.

Because of the addition of a new link, the links are moved to a
dropdown menu "Nos projets" at 1200px, then to a hamburger at 768px.

The donation button is always present on width > 576px, and is moved
to the hamburger menu below.

Closes #284