Drop number from mobilito mode icons
The icons were probably at their origin screenshots of the mobilito
starting screen, but in any case contained the count "zero" on each
icon.  This makes no sense for the tutorial.

There remains a question why we have these icons at all rather than
generating the squares the same way we do on the mobilito recording
page.  We'll leave that to another day.

Since I've edited the files, I'm adding the xcf source files as well
in this commit.

Closes #968.