Refactor mobilito to avoid weird template view inheritance
Several mobilito template views were inheriting from a common template
view class, which was really strange and difficult to understand.  It
did this in order to share code that handled tracking which tutorials
had been seen, which to view next, and when to save state for the
user.  In doing that, it ended up requiring some slighly strange
overrides to the standard get() / get_context_data() from the django
library.  Brief, it was a bit twisted and not really a natural pattern
for django.

This commit refactors the code responsible for tracking tutorial view
state to a separate class, thus simplifying the view code.  In the
process, refactor the view code to use the new tutorial state class.

Part of #966.