Add (png) visualisations of mobilito sessions
This commit adds a basic time series plot, separated by mode (in the
transportation sense: bike, ped, car).  It also adds a pie chart that
visualises the fraction bike+ped against car.

In the pocess, improve the session summary page.

This commit leaves a few opportunities for improvement in future
commits:

  * The pie chart still needs labels.  Or, rather, the labels need to
    be positioned correctly, as they are currently being drawn out of
    frame.

  * Some of the text on the session summary page is still displaying
    white instead of Mobilitains gray.

  * The mode icons might be better with dark blue background and gray
    foreground.

  * At some point we should experiment whether violin charts would
    appeal more to our user base than the jittered scatter plot.

  * The four mode summary bullet list would surely look better
    vertically centered on the pie chart.

This includes a few PIP-8 and Flak8 modifications that became too
confusing to separate into another commit.

Part of #971.